feat: データローダーの改善とフルデータ読み込み機能の実装
- データローダーをモジュール化して分割
- 全データを一度に読み込む機能を実装
- アプリケーションコントローラーとデータコントローラーを修正
