+++ cline_docs/activeContext.md
@@ -7,6 +7,7 @@
 1. **データ読み込み機能**
    - CSVファイルの読み込み
    - チャンク読み込みによる大規模データの効率的な処理
+   - 全データ一括読み込み機能
    - ヘッダー情報の解析
 
 2. **データ表示機能**
