+++ cline_docs/progress.md
@@ -18,7 +18,7 @@
 
 ### 進行中のタスク
 1. **パフォーマンス最適化**
-   - [ ] 大規模データ処理の最適化
+   - [x] 大規模データ処理の最適化
    - [ ] メモリ使用量の最適化
    - [ ] 描画パフォーマンスの改善
 
@@ -35,6 +35,7 @@
 1. **データ読み込み機能**
    - [x] CSVファイルの読み込み
    - [x] チャンク読み込みによる大規模データの効率的な処理
+   - [x] 全データ一括読み込み機能
    - [x] ヘッダー情報の解析
 
 2. **データ表示機能**
