+++ issue/10.md
@@ -1,0 +1,26 @@
+# Fix: Load all data chunks for complete graph display
+
+## 問題
+現在、データの一部（最初のチャンク）だけが読み込まれて処理されているため、グラフが本来の範囲の一部しか表示されていません。
+
+## 修正内容
+1. データローダーに全チャンクを読み込む機能を追加 ✅
+2. コントローラーで全データを読み込むように修正 ✅
+3. データプロセッサーで全データを処理するように修正 ✅
+
+## 期待される結果
+- 全範囲のデータがグラフに正しく表示される ✅
+- データの部分的な表示による誤解を防ぐ ✅
+
+## 技術的な詳細
+- BaseDataLoader クラスに load_all_data() メソッドを追加 ✅
+- 各データローダー実装クラスで load_all_data() メソッドを実装 ✅
+- AppController の _load_initial_data() メソッドを修正 ✅
+
+## 実装内容
+- データローダーをモジュール化して分割
+- 全データを一度に読み込む機能を実装
+- アプリケーションコントローラーとデータコントローラーを修正
+
+## ステータス
+✅ 完了
