+++ src/controller/app_controller.py
@@ -4,11 +4,11 @@
 アプリケーション全体の制御を担当します。
 """
 
-from .data_controller import DataController
-from .plot_controller import PlotController
-from ..model.data_loader import DataLoaderFactory
-from ..model.data_processor import DataProcessor
-from ..model.data_filter import DataFilter
+from src.controller.data_controller import DataController
+from src.controller.plot_controller import PlotController
+from src.model.data_loader import DataLoaderFactory
+from src.model.data_processor import DataProcessor
+from src.model.data_filter import DataFilter
 
 
 class AppController:
@@ -77,14 +77,14 @@ class AppController:
 
     def _load_initial_data(self):
         """初期データの読み込み"""
-        # 最初のチャンクを読み込み
-        chunk, _ = self.data_loader.get_chunk(0)
+        # すべてのデータを読み込み
+        df = self.data_loader.load_all_data()
 
         # データプロセッサーにデータを設定
-        self.data_processor.set_data(chunk)
+        self.data_processor.set_data(df)
 
         # データフィルターにデータを設定
-        self.data_filter.set_data(chunk)
+        self.data_filter.set_data(df)
 
         # 軸の設定
         x_column = self.main_window.control_panel.x_column.get()
